feat: update plugins again
