--- app/plugins/withAndroidManifestInject/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "expo-module-scripts/tsconfig.plugin",
-    "compilerOptions": {
-        "outDir": "build",
-        "rootDir": "src"
-    },
-    "include": ["./src"],
-    "exclude": ["**/__mocks__/*", "**/__tests__/*"]
-}
